allow columns to be move and resized via keyboard
