+++ grid-community-modules/core/src/ts/constants/direction.ts
@@ -1,0 +1,2 @@
+export enum VerticalDirection { Up, Down }
+export enum HorizontalDirection { Left, Right }
