+++ grid-community-modules/core/src/ts/main.ts
@@ -88,8 +88,6 @@ export {
 export {
     DragAndDropService,
     DragSourceType,
-    HorizontalDirection,
-    VerticalDirection,
     DropTarget,
     DragSource,
     DragItem,
@@ -329,6 +327,7 @@ export { VanillaFrameworkOverrides } from "./vanillaFrameworkOverrides";
 export { CellNavigationService } from "./cellNavigationService";
 export { AlignedGridsService } from "./alignedGridsService";
 export { KeyCode } from "./constants/keyCode";
+export { VerticalDirection, HorizontalDirection } from "./constants/direction";
 export { Grid, GridParams, Params, GridCoreCreator, createGrid } from "./grid";
 export { GridApi, DetailGridInfo, StartEditingCellParams } from "./gridApi";
 export { Events } from "./eventKeys";
